==> Data Analysis & Visualization/Project/data/3. transform.ipynb <==
# this file is to transform data into desired format

# import libraries

import pandas as pd
import numpy as np

# reading csv file
df = pd.read_csv('combined_data_orig.csv')
df.head()

# adding a column of date and time from year, month, day, hour in following format
# 2019-01-01 00:00:00
df['date'] = pd.to_datetime(df[['year', 'month', 'day', 'hour']]).dt.strftime('%Y-%m-%d %H:%M:%S')
df.head()

# converting date column to followin format
# July 19, 2017, 7:44 PM

df['date'] = pd.to_datetime(df['date']).dt.strftime('%B %d, %Y, %I:%M %p')
df.head()

# checking for null values
df.isnull().sum()

# drop rows with null values
df = df.dropna()

# creating a new column of smog index based on following conditions
# value equal to max from pm10 or pm2_5 if temperature is less than 25 and pm10 or pm2_5 is more than 300
# null in other case

df['smog'] = np.where((df['temperature'] < 25) & (df['pm10'] > 300), df[['pm10', 'pm2_5']].max(axis=1), np.nan)

# create a new dataframe with following conditions
# year = 2022
# month = 11
df_test = df[(df['year'] == 2022) & (df['month'] == 11)]

# drop rows with year = 2022 and month = 11
df = df.drop(df_test.index)

# # create a new dataframe with following conditions
# # year = 2022
# # month = 10
# df_test = df[(df['year'] == 2022) & (df['month'] == 10)]

# # drop rows with year = 2022 and month = 11
# df = df.drop(df_test.index)

# storing the data in csv file
df.to_csv('combined_data.csv', index=False)
df_test.to_csv('test_data.csv', index=False)
