==> Data Analysis & Visualization/Project/data/4. single_city.ipynb <==
# this file extracts records for a single city from the original dataset
# and saves it to a new file

import pandas as pd

data=pd.read_csv('combined_data.csv')
data.head()

# selecting only data for Islamabad
data=data[data['city']=='Islamabad']
data.head()

data.to_csv('islamabad.csv',index=False)
